renamed to scanimageimaginginterfaces.py
