+++ src/neuroconv/datainterfaces/ophys/scanimage/scanimageimaginginterfaces.py
